Backtester migrated from the private repo

+++ common/Common.Interfaces.Api/StaticData/ContractsFilter.cs
@@ -3,7 +3,7 @@ namespace QuantInfra.Common.Interfaces.Api.StaticData;
 public class ContractsFilter : PagingFilter
 {
     public virtual string? Ticker { get; set; }
-    public virtual long? ExchangeId { get; set; } = null;
+    public virtual int? ExchangeId { get; set; } = null;
     public virtual List<int>? ContractIds { get; set; } = null;
-    public virtual long? CommissionId { get; set; }
+    public virtual int? CommissionId { get; set; }
 }
